Remove intermediate class BaseRadialAcceptanceMapCreator and move non-common methods out of BaseAcceptanceMapCreator
